First working but not dynamic version of table.
