+++ src/_pytask/collect_command.py
@@ -11,11 +11,11 @@ from _pytask.enums import ColorCode
 from _pytask.enums import ExitCode
 from _pytask.exceptions import CollectionError
 from _pytask.exceptions import ConfigurationError
-from _pytask.nodes import reduce_node_name
 from _pytask.path import find_common_ancestor
 from _pytask.path import relative_to
 from _pytask.pluginmanager import get_plugin_manager
 from _pytask.session import Session
+from _pytask.shared import reduce_node_name
 from rich.tree import Tree
