+++ src/_pytask/nodes.py
@@ -14,14 +14,9 @@ from typing import Tuple
 from typing import Union
 
 import attr
-from _pytask.console import escape_squared_brackets
 from _pytask.exceptions import NodeNotCollectedError
 from _pytask.exceptions import NodeNotFoundError
-from _pytask.mark import get_marks_from_obj
-from _pytask.path import find_closest_ancestor
-from _pytask.path import find_common_ancestor
-from _pytask.path import relative_to
-from _pytask.shared import find_duplicates
+from _pytask.mark_utils import get_marks_from_obj
 
 
 def depends_on(objects: Union[Any, Iterable[Any]]) -> Union[Any, Iterable[Any]]:
@@ -371,38 +366,23 @@ def create_task_name(path: Path, base_name: str):
     return path.as_posix() + "::" + base_name
 
 
-def reduce_node_name(node, paths: List[Path]):
-    """Reduce the node name.
+def find_duplicates(x: Iterable[Any]):
+    """Find duplicated entries in iterable.
 
-    The whole name of the node - which includes the drive letter - can be very long
-    when using nested folder structures in bigger projects.
-
-    Thus, the part of the name which contains the path is replaced by the relative
-    path from one path in ``session.config["paths"]`` to the node.
+    Examples
+    --------
+    >>> find_duplicates(["a", "b", "a"])
+    {'a'}
+    >>> find_duplicates(["a", "b"])
+    set()
 
     """
-    ancestor = find_closest_ancestor(node.path, paths)
-    if ancestor is None:
-        try:
-            ancestor = find_common_ancestor(node.path, *paths)
-        except ValueError:
-            ancestor = node.path.parents[-1]
+    seen = set()
+    duplicates = set()
 
-    if isinstance(node, MetaTask):
-        shortened_path = relative_to(node.path, ancestor)
-        raw_name = create_task_name(shortened_path, node.base_name)
-        name = escape_squared_brackets(raw_name)
-    elif isinstance(node, MetaNode):
-        name = relative_to(node.path, ancestor).as_posix()
-    else:
-        raise ValueError(f"Unknown node {node} with type '{type(node)}'.")
+    for i in x:
+        if i in seen:
+            duplicates.add(i)
+        seen.add(i)
 
-    return name
-
-
-def reduce_names_of_multiple_nodes(names, dag, paths):
-    """Reduce the names of multiple nodes in the DAG."""
-    return [
-        reduce_node_name(dag.nodes[n].get("node") or dag.nodes[n].get("task"), paths)
-        for n in names
-    ]
+    return duplicates
