+++ src/_pytask/parametrize.py
@@ -15,7 +15,7 @@ from _pytask.console import format_strings_as_flat_tree
 from _pytask.console import TASK_ICON
 from _pytask.mark import Mark
 from _pytask.mark import MARK_GEN as mark  # noqa: N811
-from _pytask.shared import find_duplicates
+from _pytask.nodes import find_duplicates
 
 
 def parametrize(
@@ -129,7 +129,7 @@ def pytask_parametrize_task(session, name, obj):
             )
             raise ValueError(
                 "The following ids are duplicated while parametrizing task "
-                f"'{obj.__name__}'.\n\n{text}\n\nIt might be caused by "
+                f"'{name}'.\n\n{text}\n\nIt might be caused by "
                 "parametrizing the task with the same combination of arguments "
                 "multiple times. Change the arguments or change the ids generated by "
                 "the parametrization."
