+++ tests/test_nodes.py
@@ -16,7 +16,7 @@ from _pytask.nodes import MetaNode
 from _pytask.nodes import MetaTask
 from _pytask.nodes import produces
 from _pytask.nodes import PythonFunctionTask
-from _pytask.nodes import reduce_node_name
+from _pytask.shared import reduce_node_name
 
 
 @pytest.mark.unit
